requirements.txt
pytest>=4.4
psycopg2-binary>=2.8
pandas>=0.24
boto3>=1.9
python-dotenv>=0.10
ipykernel>=5.1
jupyterlab>=0.35
botocore>=1.12
pylint>=2.3
matplotlib>=3.0
